se cambiaron los tipos de datos de las tabals

+++ database/migrations/2025_11_12_171031_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');  
             $table->text('description')->nullable();
-            $table->decimal('price',10,2);
+            $table->float('price',10,2);
             $table->string('photo')->nullable();
             $table->timestamps();
         });
